requirements.txt
absl-py==0.6.1
altgraph==0.16.1
astor==0.7.1
cycler==0.10.0
future==0.17.1
gast==0.2.0
grpcio==1.16.1
h5py==2.8.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
macholib==1.11
Markdown==3.0.1
matplotlib==3.0.2
numpy==1.15.4
pefile==2019.4.18
protobuf==3.6.1
pydot==1.3.0
PyInstaller==3.4
pyparsing==2.3.0
python-dateutil==2.7.5
PyYAML==3.13
scipy==1.1.0
SimpleWebSocketServer==0.1.1
six==1.11.0
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
websocket-client==0.56.0
Werkzeug==0.14.1
yappi==0.99
